Organized files, fixed bugs,  and allow popover to be accessible to other plugins
