+++ public/hacks/custom_filter_bootstrap/FilterManagerHelper.js
